schema: rename test package to schematest
