+++ schema/epp/response_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/domainr/epp2/schema/epp"
+	"github.com/domainr/epp2/schema/schematest"
 	"github.com/domainr/epp2/schema/std"
-	"github.com/domainr/epp2/schema/test"
 )
 
 func TestResponseRoundTrip(t *testing.T) {
@@ -135,7 +135,7 @@ func TestResponseRoundTrip(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			test.RoundTrip(t, nil, tt.v, tt.want, tt.wantErr)
+			schematest.RoundTrip(t, nil, tt.v, tt.want, tt.wantErr)
 		})
 	}
 }
